requirements.txt
streamlit
torch==2.7.1
-f https://download.pytorch.org/whl/torch_stable.html
openai-whisper
deep-translator
opencv-python
srt
numpy
pillow
gunicorn
pymongo[srv]==4.6.3
dnspython==2.4.2
certifi
bcrypt